requirements.txt
ccxt==4.4.45
python-dotenv==1.0.1
pytz==2024.1
